backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
google-generativeai==0.8.3
python-dotenv==1.0.0
marshmallow==3.20.1
protobuf>=5.29.0
gunicorn==21.2.0
